Add formatting and anyhow

+++ src/cmd/remove.rs
@@ -1,5 +1,7 @@
+use anyhow::anyhow;
 use std::{fs, path::Path};
 
+use anyhow::Result;
 use clap::{value_parser, Arg, ArgMatches, Command};
 
 use crate::kubeconfig;
@@ -14,19 +16,17 @@ pub fn command() -> Command {
         .arg_required_else_help(true)
 }
 
-pub fn execute(config_path: &Path, matches: &ArgMatches) -> Result<(), Box<dyn std::error::Error>> {
+pub fn execute(config_path: &Path, matches: &ArgMatches) -> Result<()> {
     let config = matches
         .get_one::<String>("kubeconfig")
-        .ok_or("failed to get kubeconfig argument")?;
+        .ok_or_else(|| anyhow!("failed to get kubeconfig argument"))?;
 
     let kubeconfig_path = config_path.join(format!("{config}.kubeconfig"));
 
-    match kubeconfig::get(&kubeconfig_path) {
-        Ok(_) => {
-            fs::remove_file(&kubeconfig_path)?;
-            log::info!("removed kubeconfig at {}", kubeconfig_path.display());
-            Ok(())
-        }
-        Err(err) => Err(err),
+    if kubeconfig::get(&kubeconfig_path).is_ok() {
+        fs::remove_file(&kubeconfig_path)?;
+        log::info!("removed kubeconfig at {}", kubeconfig_path.display());
+        return Ok(());
     }
+    Err(anyhow!("Kubeconfig not found: {:?}", kubeconfig_path))
 }
